# lib/api-zod/src/generated/types/paidMediaError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NexOS AI — Campaign execution and launch automation API
 * OpenAPI spec version: 0.1.0
 */

export interface PaidMediaError {
  error: string;
  code: string;
}