# lib/api-zod/src/generated/types/paidMediaPoliciesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NexOS AI — Campaign execution and launch automation API
 * OpenAPI spec version: 0.1.0
 */
import type { PaidMediaPolicy } from "./paidMediaPolicy";

export interface PaidMediaPoliciesResponse {
  policies: PaidMediaPolicy[];
}