# lib/api-zod/src/generated/types/paidMediaProposal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NexOS AI — Campaign execution and launch automation API
 * OpenAPI spec version: 0.1.0
 */
import type { PaidMediaProposalActionType } from "./paidMediaProposalActionType";
import type { PaidMediaProposalAfterAllocation } from "./paidMediaProposalAfterAllocation";
import type { PaidMediaProposalBeforeAllocation } from "./paidMediaProposalBeforeAllocation";
import type { PaidMediaProposalMetrics } from "./paidMediaProposalMetrics";
import type { PaidMediaProposalPolicyDecision } from "./paidMediaProposalPolicyDecision";
import type { PaidMediaProposalProvider } from "./paidMediaProposalProvider";
import type { PaidMediaProposalRequestedChange } from "./paidMediaProposalRequestedChange";
import type { PaidMediaProposalSimulation } from "./paidMediaProposalSimulation";
import type { PaidMediaProposalStatus } from "./paidMediaProposalStatus";

export interface PaidMediaProposal {
  id: string;
  workspaceId: string;
  /** @nullable */
  accountId?: string | null;
  /** @nullable */
  entityId?: string | null;
  provider: PaidMediaProposalProvider;
  actionType: PaidMediaProposalActionType;
  status: PaidMediaProposalStatus;
  idempotencyKey: string;
  recommendation: string;
  metrics: PaidMediaProposalMetrics;
  simulation: PaidMediaProposalSimulation;
  beforeAllocation?: PaidMediaProposalBeforeAllocation;
  afterAllocation?: PaidMediaProposalAfterAllocation;
  requestedChange?: PaidMediaProposalRequestedChange;
  policyDecision?: PaidMediaProposalPolicyDecision;
  expiresAt: Date;
}